+++ docs/README.md
@@ -50,8 +50,7 @@
 ### 플레이어의 수익 정책
 #### 플레이어가 이긴 경우
 - [x] 플레이어가 BlackJack으로 승리한 경우, 배팅 금액의 1.5배를 준다.
-- [x] 딜러가 버스트여서 플레이어가 승리한 경우, 배팅 금액의 0배를 준다. (= 추가 수익이 없다.)
-- [x] 딜러가 버스트가 아닌데도 플레이어가 승리한 경우, 배팅 금액의 1배를 준다.
+- [x] 플레이어가 승리한 경우, 배팅 금액의 1배를 준다.
 
 #### 플레이어가 무승부인 경우
 - [x] 플레이어가 무승부인 경우, 배팅 금액의 0배를 준다.
