+++ src/main/java/blackjack/model/betting/BettingRule.java
@@ -7,16 +7,10 @@ import static blackjack.model.betting.ProfitRate.ZERO;
 import static blackjack.model.result.ResultCommand.DRAW;
 import static blackjack.model.result.ResultCommand.WIN;
 
-import blackjack.model.participant.Dealer;
 import blackjack.model.participant.Player;
 import blackjack.model.result.ResultCommand;
 
 public class BettingRule {
-    private final Dealer dealer;
-
-    public BettingRule(final Dealer dealer) {
-        this.dealer = dealer;
-    }
 
     public ProfitRate calculateProfitRate(final Player player, final ResultCommand resultCommand) {
         if (resultCommand.equals(WIN)) {
@@ -29,10 +23,6 @@ public class BettingRule {
     private ProfitRate calculateProfitRateWhenPlayerWin(final Player player) {
         if (player.isBlackJack()) {
             return PLUS_150_PERCENT;
-        }
-
-        if (dealer.isBust()) {
-            return ZERO;
         }
 
         return PLUS_100_PERCENT;
